// lib/api/generated/models/createApiKeyResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Heymoa API
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKeyStatus } from "./apiKeyStatus";

export interface CreateApiKeyResponse {
  id: string;
  name: string;
  secretKey: string;
  maskedKey: string;
  status: ApiKeyStatus;
  createdAt: string;
}